# Simplify mergedPostings.Seek